Merge from master

+++ gpu4pyscf/lib/gint/g2e_root3.cu
@@ -206,9 +206,9 @@ static void GINTfill_int2e_kernel0022(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -464,9 +464,9 @@ static void GINTfill_int2e_kernel0031(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -782,9 +782,9 @@ static void GINTfill_int2e_kernel0032(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -1155,9 +1155,9 @@ static void GINTfill_int2e_kernel1021(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -1642,9 +1642,9 @@ static void GINTfill_int2e_kernel1022(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -2048,9 +2048,9 @@ static void GINTfill_int2e_kernel1030(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -2445,9 +2445,9 @@ static void GINTfill_int2e_kernel1031(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -2948,9 +2948,9 @@ static void GINTfill_int2e_kernel1111(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -3364,9 +3364,9 @@ static void GINTfill_int2e_kernel1120(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -3981,9 +3981,9 @@ static void GINTfill_int2e_kernel1121(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -4643,9 +4643,9 @@ static void GINTfill_int2e_kernel1130(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -5077,9 +5077,9 @@ static void GINTfill_int2e_kernel2011(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -5391,9 +5391,9 @@ static void GINTfill_int2e_kernel2020(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -5843,9 +5843,9 @@ static void GINTfill_int2e_kernel2021(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -6322,9 +6322,9 @@ static void GINTfill_int2e_kernel2030(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -6695,9 +6695,9 @@ static void GINTfill_int2e_kernel2110(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -7312,9 +7312,9 @@ static void GINTfill_int2e_kernel2111(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -8016,9 +8016,9 @@ static void GINTfill_int2e_kernel2120(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -8433,9 +8433,9 @@ static void GINTfill_int2e_kernel2200(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -8884,9 +8884,9 @@ static void GINTfill_int2e_kernel2210(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -9290,9 +9290,9 @@ static void GINTfill_int2e_kernel3010(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -9688,9 +9688,9 @@ static void GINTfill_int2e_kernel3011(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -10131,9 +10131,9 @@ static void GINTfill_int2e_kernel3020(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -10437,9 +10437,9 @@ static void GINTfill_int2e_kernel3100(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j; 
@@ -10834,9 +10834,9 @@ static void GINTfill_int2e_kernel3110(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
@@ -11272,9 +11272,9 @@ static void GINTfill_int2e_kernel3200(GINTEnvVars envs, ERITensor eri, BasisProd
         }
     } }
 
-    size_t jstride = eri.stride_j;
-    size_t kstride = eri.stride_k;
-    size_t lstride = eri.stride_l;
+    int jstride = eri.stride_j;
+    int kstride = eri.stride_k;
+    int lstride = eri.stride_l;
     int *ao_loc = c_bpcache.ao_loc;
     int i0 = ao_loc[ish] - eri.ao_offsets_i;
     int j0 = ao_loc[jsh] - eri.ao_offsets_j;
